requirements.txt
line-bot-sdk
fastapi
python-dotenv
uvicorn[standard]
google-cloud-language
wheel